# Optimizing ByChained#findElement