--- a/common/src/web/byChainedTest.html
+++ b/common/src/web/byChainedTest.html
@@ -1,0 +1,19 @@
+<?xml version="1.0"?>
+<html xmlns="http://www.w3.org/1999/xhtml" xml:lang="en" lang="en">
+<head>
+    <title>ByChained Test Page</title>
+</head>
+<body>
+
+<div id=nav">
+  <div id="elem1" name="cheese">
+    <div id="elem3" name="photo"></div>
+  </div>
+</div>
+
+<div id="elem2" name="cheese">
+  <div id="elem5" name="click"></div>
+</div>
+
+</body>
+</html>