--- a/java/client/test/org/openqa/selenium/support/LargeTests.java
+++ b/java/client/test/org/openqa/selenium/support/LargeTests.java
@@ -19,6 +19,7 @@ package org.openqa.selenium.support;
 import org.junit.AfterClass;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
+import org.openqa.selenium.support.pagefactory.ByChainedIntegrationTest;
 import org.openqa.selenium.support.pagefactory.UsingPageFactoryTest;
 import org.openqa.selenium.support.ui.SelectElementTest;
 import org.openqa.selenium.support.ui.SelectLargeTest;
@@ -28,7 +29,8 @@ import org.openqa.selenium.testing.JUnit4TestBase;
 @Suite.SuiteClasses({
     SelectLargeTest.class,
     UsingPageFactoryTest.class,
-    SelectElementTest.class
+    SelectElementTest.class,
+    ByChainedIntegrationTest.class
 })
 public class LargeTests {