- optimize complexity: instead of traversing the entire "tree"
  of partial WebElements found, only travers the left-most path
- in case the bys list is empty, throw an exception with a more meaningful message

Not sure if you like the

```
assert elem != null;
```

I seem to have some issues in running the tests. They are failing, even on master. I guess it must be due to my env (?).
OSX 10.10.2, Java 1.8.0_40
Also tried java 6 but it didnt seem to help

```
Compiling: //java/client/test/org/openqa/selenium:SmallTests as build/java/client/test/org/openqa/selenium/SmallTests.jar
mkdir -p build/java/client/test/org/openqa/selenium/SmallTests.jar_temp
Compiling 1 source file to /Users/alb/workspace/selenium/build/java/client/test/org/openqa/selenium/SmallTests.jar_temp
Building jar: /Users/alb/workspace/selenium/build/java/client/test/org/openqa/selenium/SmallTests.jar
rm -rf build/java/client/test/org/openqa/selenium/SmallTests.jar_temp
** Execute //java/client/test/org/openqa/selenium:SmallTests
** Execute //java/client/test/org/openqa/selenium:SmallTests:run
Testing: //java/client/test/org/openqa/selenium:SmallTests
mkdir -p build/test_logs
Listening for transport dt_socket at address: 5005
Running org.openqa.selenium.SmallTests
Exception in thread "Thread-20" org.openqa.selenium.WebDriverException: Thread safety error; this instance of WebDriver was constructed on thread main (id 1) and is being accessed by thread Thread-20 (id 35)This is not permitted and *will* cause undefined behaviour
Build info: version: '2.45.0', revision: '17dc63a', time: '2015-03-26 03:09:18'
System info: host: 'albs-MacBook-Air.local', ip: '192.168.0.25', os.name: 'Mac OS X', os.arch: 'x86_64', os.version: '10.10.2', java.version: '1.8.0_40'
Driver info: driver.version: unknown
    at org.openqa.selenium.support.ThreadGuard$WebDriverInvocationHandler.invoke(ThreadGuard.java:86)
    at com.sun.proxy.$Proxy22.findElement(Unknown Source)
    at org.openqa.selenium.support.ThreadGuardTest$1.run(ThreadGuardTest.java:44)
    at java.lang.Thread.run(Thread.java:745)
Tests run: 345, Failures: 0, Errors: 0, Time elapsed: 1.831 sec
** Invoke test_support (first_time)
** Invoke //java/client/test/org/openqa/selenium/lift:test:run (first_time)
** Invoke //java/client/test/org/openqa/selenium/lift:test
** Execute //java/client/test/org/openqa/selenium/lift:test:run
Testing: //java/client/test/org/openqa/selenium/lift:test
mkdir -p build/test_logs
Listening for transport dt_socket at address: 5005
Running org.openqa.selenium.lift.SmallTests
Tests run: 23, Failures: 0, Errors: 0, Time elapsed: 0.394 sec
** Invoke //java/client/test/org/openqa/selenium/support:SmallTests:run (first_time)
** Invoke //java/client/test/org/openqa/selenium/support:SmallTests (first_time)
** Invoke build/java/client/test/org/openqa/selenium/support/SmallTests.jar (first_time)
** Invoke //java/client/test/org/openqa/selenium/support:tests
** Invoke java/client/test/org/openqa/selenium/support/SmallTests.java (not_needed)
** Execute build/java/client/test/org/openqa/selenium/support/SmallTests.jar
rm -rf build/java/client/test/org/openqa/selenium/support/SmallTests.jar_temp
Exception `Errno::ENOENT' at org/jruby/RubyFile.java:1729 - No such file or directory - build/java/client/test/org/openqa/selenium/support/SmallTests.jar_temp
Compiling: //java/client/test/org/openqa/selenium/support:SmallTests as build/java/client/test/org/openqa/selenium/support/SmallTests.jar
mkdir -p build/java/client/test/org/openqa/selenium/support/SmallTests.jar_temp
Compiling 1 source file to /Users/alb/workspace/selenium/build/java/client/test/org/openqa/selenium/support/SmallTests.jar_temp
Building jar: /Users/alb/workspace/selenium/build/java/client/test/org/openqa/selenium/support/SmallTests.jar
rm -rf build/java/client/test/org/openqa/selenium/support/SmallTests.jar_temp
** Execute //java/client/test/org/openqa/selenium/support:SmallTests
** Execute //java/client/test/org/openqa/selenium/support:SmallTests:run
Testing: //java/client/test/org/openqa/selenium/support:SmallTests
mkdir -p build/test_logs
Listening for transport dt_socket at address: 5005
Running org.openqa.selenium.support.SmallTests
Tests run: 181, Failures: 0, Errors: 0, Time elapsed: 1.158 sec
Exception in thread "Thread-1" org.openqa.selenium.WebDriverException: Thread safety error; this instance of WebDriver was constructed on thread main (id 1) and is being accessed by thread Thread-1 (id 12)This is not permitted and *will* cause undefined behaviour
Build info: version: '2.45.0', revision: '17dc63a', time: '2015-03-26 03:09:18'
System info: host: 'albs-MacBook-Air.local', ip: '192.168.0.25', os.name: 'Mac OS X', os.arch: 'x86_64', os.version: '10.10.2', java.version: '1.8.0_40'
Driver info: driver.version: unknown
    at org.openqa.selenium.support.ThreadGuard$WebDriverInvocationHandler.invoke(ThreadGuard.java:86)
    at com.sun.proxy.$Proxy20.findElement(Unknown Source)
    at org.openqa.selenium.support.ThreadGuardTest$1.run(ThreadGuardTest.java:44)
    at java.lang.Thread.run(Thread.java:745)
** Invoke //java/client/test/org/openqa/selenium/support:LargeTests:run (first_time)
** Invoke //java/client/test/org/openqa/selenium/support:LargeTests (first_time)
** Invoke build/java/client/test/org/openqa/selenium/support/LargeTests.jar (first_time)
** Invoke //java/client/test/org/openqa/selenium/support:tests
** Invoke java/client/test/org/openqa/selenium/support/LargeTests.java (not_needed)
** Execute build/java/client/test/org/openqa/selenium/support/LargeTests.jar
rm -rf build/java/client/test/org/openqa/selenium/support/LargeTests.jar_temp
Exception `Errno::ENOENT' at org/jruby/RubyFile.java:1729 - No such file or directory - build/java/client/test/org/openqa/selenium/support/LargeTests.jar_temp
Compiling: //java/client/test/org/openqa/selenium/support:LargeTests as build/java/client/test/org/openqa/selenium/support/LargeTests.jar
mkdir -p build/java/client/test/org/openqa/selenium/support/LargeTests.jar_temp
Compiling 1 source file to /Users/alb/workspace/selenium/build/java/client/test/org/openqa/selenium/support/LargeTests.jar_temp
Building jar: /Users/alb/workspace/selenium/build/java/client/test/org/openqa/selenium/support/LargeTests.jar
rm -rf build/java/client/test/org/openqa/selenium/support/LargeTests.jar_temp
** Execute //java/client/test/org/openqa/selenium/support:LargeTests
** Execute //java/client/test/org/openqa/selenium/support:LargeTests:run
Testing: //java/client/test/org/openqa/selenium/support:LargeTests
mkdir -p build/test_logs
Listening for transport dt_socket at address: 5005
Running org.openqa.selenium.support.LargeTests
Mar 26, 2015 3:12:14 AM org.openqa.selenium.testing.JUnit4TestBase$1 starting
INFO: >>> Starting selectByVisibleTextShouldNormalizeSpaces(org.openqa.selenium.support.ui.SelectLargeTest)
2015-03-26 03:12:14.709:INFO:osjs.Server:jetty-7.x.y-SNAPSHOT
2015-03-26 03:12:14.744:INFO:osjsh.ContextHandler:started o.s.j.s.ServletContextHandler{/common,file:/Users/alb/workspace/selenium/common/src/web/}
2015-03-26 03:12:14.748:INFO:/common:Aliases are enabled
2015-03-26 03:12:14.749:INFO:osjsh.ContextHandler:started o.s.j.s.ServletContextHandler{/javascript,file:/Users/alb/workspace/selenium/javascript/}
2015-03-26 03:12:14.751:INFO:/javascript:Aliases are enabled
2015-03-26 03:12:14.752:INFO:osjsh.ContextHandler:started o.s.j.s.ServletContextHandler{/third_party/closure/goog,file:/Users/alb/workspace/selenium/third_party/closure/goog/}
2015-03-26 03:12:14.752:INFO:/third_party/closure/goog:Aliases are enabled
2015-03-26 03:12:14.752:INFO:osjsh.ContextHandler:started o.s.j.s.ServletContextHandler{/third_party/js,file:/Users/alb/workspace/selenium/third_party/js/}
2015-03-26 03:12:14.752:INFO:/third_party/js:Aliases are enabled
2015-03-26 03:12:14.778:INFO:osjs.AbstractConnector:Started SelectChannelConnector@0.0.0.0:34396
2015-03-26 03:12:14.996:INFO:osjus.SslContextFactory:Enabled Protocols [SSLv2Hello, TLSv1, TLSv1.1, TLSv1.2] of [SSLv2Hello, SSLv3, TLSv1, TLSv1.1, TLSv1.2]
2015-03-26 03:12:15.007:INFO:osjs.AbstractConnector:Started SslSocketConnector@0.0.0.0:20733
Tests run: 5, Failures: 0, Errors: 5, Time elapsed: 0.757 sec
Mar 26, 2015 3:12:15 AM org.openqa.selenium.testing.JUnit4TestBase$1 finished
INFO: <<< Finished selectByVisibleTextShouldNormalizeSpaces(org.openqa.selenium.support.ui.SelectLargeTest)
Mar 26, 2015 3:12:15 AM org.openqa.selenium.testing.JUnit4TestBase$1 starting
INFO: >>> Starting multipleSelectShouldBePossibleIfMulitpleAttributeEmpty(org.openqa.selenium.support.ui.SelectLargeTest)
Mar 26, 2015 3:12:15 AM org.openqa.selenium.testing.JUnit4TestBase$1 finished
INFO: <<< Finished multipleSelectShouldBePossibleIfMulitpleAttributeEmpty(org.openqa.selenium.support.ui.SelectLargeTest)
Mar 26, 2015 3:12:15 AM org.openqa.selenium.testing.JUnit4TestBase$1 starting
INFO: >>> Starting canListDecoratedElements(org.openqa.selenium.support.pagefactory.UsingPageFactoryTest)
Mar 26, 2015 3:12:15 AM org.openqa.selenium.testing.JUnit4TestBase$1 finished
INFO: <<< Finished canListDecoratedElements(org.openqa.selenium.support.pagefactory.UsingPageFactoryTest)
Mar 26, 2015 3:12:15 AM org.openqa.selenium.testing.JUnit4TestBase$1 starting
INFO: >>> Starting testDecoratedElementsShouldBeUnwrapped(org.openqa.selenium.support.pagefactory.UsingPageFactoryTest)
Mar 26, 2015 3:12:15 AM org.openqa.selenium.testing.JUnit4TestBase$1 finished
INFO: <<< Finished testDecoratedElementsShouldBeUnwrapped(org.openqa.selenium.support.pagefactory.UsingPageFactoryTest)
Mar 26, 2015 3:12:15 AM org.openqa.selenium.testing.JUnit4TestBase$1 starting
INFO: >>> Starting canExecuteJsUsingDecoratedElements(org.openqa.selenium.support.pagefactory.UsingPageFactoryTest)
Mar 26, 2015 3:12:15 AM org.openqa.selenium.testing.JUnit4TestBase$1 finished
INFO: <<< Finished canExecuteJsUsingDecoratedElements(org.openqa.selenium.support.pagefactory.UsingPageFactoryTest)
go aborted!
org.apache.tools.ant.BuildException: Test org.openqa.selenium.support.LargeTests failed
```