@lukeis 2015-03-26
regarding your local build, not sure what's going on there. But you should be able to use Java 8 or 7 to build / run (but not 6, definitely not 6 ;) )