@alb-i986 2015-03-26
About the build, maybe I will try to setup a Linux VM on my Mac.
By the way it would be cool if you could provide a Vagrantfile for hassle free building. If you like the idea I may work on that. I already have something from which to start: see https://github.com/alb-i986/vagrantfiles/tree/master/jenkins