@lukeis 2015-03-26
You should be able to build Selenium on all platforms (windows, mac, linux). But I did start the effort there using docker once... got it "working" but since we aren't doing native events for firefox anymore, stopped working on it.