@p0deje 2015-07-08
@alb-i986 I have started working on the same thing though it only supports Firefox at the moment (which was fine for me at that point). Take a look at https://github.com/p0deje/selenium-boxes.