@alb-i986 2015-07-08
@p0deje 
I'm now finding travis a very good way to run tests hassle-free.
BTW in my case the build (the tests) was failing because I didn't have Firefox installed (LOL).