@alb-i986 2015-07-08
Ok I've just pushed my new changes to this branch.
I made quite a lot of commits, before merging I'll squash.
Thanks