@alb-i986 2015-07-09
Darn, I rebased on top of the current master, and looks like the comments made over my original commit got lost.
I expected github to keep the comments despite the different hash..
My apologies

Anyways, I'll try to sum up.

So, the discussion was revolving around the original behavior of findElement
I was saying that to me it looks like the original behavior of `ByChained#findElements` is to return an empty list if *any* one of the bys is empty. I've added a couple of unit tests which prove it.

The key is the assignment `elems = newElems` at the end of each iteration of the outer loop. If any one of the bys finds no elements, then `newElems` will be empty, thus the next iteration `elems` will be empty, and so it will always be until the end of the loop, as the inner loop will not be executed.
This BTW lead me to refactor `findElements` so that it exits the outer loop as soon as the inner loop yields no elements.

Then I was wondering if this refactor to `findElement` is really an improvement.
The question should be whether `DriverCommand.FIND_ELEMENT` is actually more efficient than `DriverCommand.FIND_ELEMENTS`.
My concern being that `DriverCommand.FIND_ELEMENT`'s might be implemented in terms of `DriverCommand.FIND_ELEMENTS`. I'd expect not to, but I don't know.

Hope not to have forgotten anything.
Cheers