@alb-i986 2016-07-12
Hey @lukeis 

I've completely re-written this PR.
As you said, it was flawed. At the time I didn't understand what you were trying to say, but after writing and debugging a couple of integration tests, now I do.
I've implemented a recursive algorithm which scans the entire tree until it finds the deepest left-most element.
Also, I took out many of the extra changes I introduced in the PR originally, which added noise and confusion.
Should be fine now.
Please check it out.
Thanks