@diemol 2020-05-20
Apologies for the poor handling of this PR.

In general, we are discouraging the use of `PageFactory`, but in any case, lots of people use it. We don't want to make more changes to it, and in any case, changes must ensure backward compatibility is complicated. 

Closing this PR.