@alb-i986 2020-05-20
Hey,
> In general, we are discouraging the use of PageFactory
Yeah? Is it reported in the docs?
Thanks