@alb-i986 2020-05-21
https://www.selenium.dev/documentation/en/guidelines_and_recommendations/page_object_models/

Here PageFactory is mentioned but not discouraged:

> There are other design patterns that also may be used in testing. Some use a Page Factory for instantiating their page objects [..]