@diemol 2020-05-21
That's a good point @alb-i986! We'll add a note there soon.