@barancev 2020-05-21
I think that this PR is safe enough from backward compatibility perspective, it does not break the API and improves internals of PageFactory implementation.