@alb-i986 2020-05-28
I'm working on this but I'm having troubles with getting the build working.
I've tried requesting for help on IRC but I didn't get any reply.
Hopefully you can help me.
I'm on Ubuntu 18.04.4. I got as far as getting a complaint about the "chalk" node module.
Here are the logs: pastebin.com/hkqGM7gC pastebin.com/FXSfGjkS
Thanks.