@alb-i986 2020-05-30
I've just created a new patch #8370 which addresses the doubts which arose here, around the functionality of the ByChained. If you look at the review comments above, you can see the original implementation was pretty obscure.
I thought it made sense to open a different PR and address that first, so as to make things clear.
Please check it out.